+++ main/src/main/java/net/oneandone/stool/stage/SourceStage.java
@@ -27,6 +27,7 @@ import org.apache.maven.project.MavenProject;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -68,6 +69,17 @@ public class SourceStage extends Stage {
     @Override
     public int size() throws IOException {
         return wars().size();
+    }
+
+    @Override
+    public List<String> vhostNames() throws IOException {
+        List<String> names;
+
+        names = new ArrayList<>();
+        for (MavenProject project : wars()) {
+            names.add(project.getArtifactId());
+        }
+        return names;
     }
 
     @Override